+++ test/core/input_reg.sv
@@ -1,0 +1,6 @@
+module Example(
+    input reg inp,
+    output reg out
+);
+    assign out = ~inp;
+endmodule