+++ test/core/input_reg.v
@@ -1,0 +1,6 @@
+module Example(
+    input wire inp,
+    output wire out
+);
+    assign out = ~inp;
+endmodule